IOC builder: Remove fragile autosuper functionality.
The meta-class support.autosuper_meta no longer supports __super on classes.
The problem is that this functionality is fragile, extremely difficult to
understand, and breaks a practical use of the builder.

Unfortunately builder support in the following modules will need to be 
updated: ipac, streamDevice, streamDevice_v1.  This is a matter of changing
    cls.__super  to  super(<class-name>, cls)
where this occurs.


git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/iocbuilder@41504 e099a375-04f9-0310-9d5f-a741eaff62e1