IOC builder: Refine mechanism for specifying EPICS_BASE.
The default EPICS_BASE is now picked up from the environment, and
can now be overridden by a Configure...() call, and furthermore
there no longer is a default if EPICS_BASE is not set.  Also the
misnamed DLS_EPICS_VERSION symbol is no longer inspected, instead
EPICS_BASE is required to conform to the DLS standard format.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/iocbuilder@43995 e099a375-04f9-0310-9d5f-a741eaff62e1