IOC builder: Rework comment policy to support doxygen documentation.
Implement a new commenting policy.  First of all, all docstrings are
removed except for module docstrings.  This works with doxygen (which
is, unfortunately, very quirky).  This involves moving all docstrings
into comments.

Secondly, we differentiate between methods which are documented, which
may be used by people writing module builder support, and internal
methods which are not included in the doxygen build.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/iocbuilder@44247 e099a375-04f9-0310-9d5f-a741eaff62e1